+++ src/index.js
@@ -1,6 +1,11 @@
 export default {
   endOfLine: 'auto',
-  plugins: ['prettier-plugin-organize-imports', 'prettier-plugin-css-order', 'prettier-plugin-packagejson'],
+  plugins: [
+    'prettier-plugin-jsdoc',
+    'prettier-plugin-organize-imports',
+    'prettier-plugin-css-order',
+    'prettier-plugin-packagejson',
+  ],
   printWidth: 120,
   quoteProps: 'consistent',
   semi: false,